Created separate class to store guild preferences [#78]
