+++ commands/random.js
@@ -1,12 +1,12 @@
 const { startGame } = require("../helpers/utils.js");
 
 module.exports = {
-    call: ({ message, db, gameSession }) => {
+    call: ({ message, db, gameSession, guildPreference }) => {
         if (!message.member.voiceChannel) {
             message.channel.send("Send `!random` again when you are in a voice channel.");
         }
         else {
-            startGame(gameSession, db, message);
+            startGame(gameSession, guildPreference, db, message);
         }
     }
 }
