+++ helpers/utils.js
@@ -37,7 +37,7 @@ module.exports = {
             return;
         }
     },
-    startGame: (gameSession, db, message) => {
+    startGame: (gameSession, guildPreference, db, message) => {
         if (gameSession.gameInSession()) {
             message.channel.send({
                 embed: {
@@ -50,11 +50,11 @@ module.exports = {
         let query = `SELECT nome as name, name as artist, vlink as youtubeLink FROM app_kpop INNER JOIN app_kpop_group ON app_kpop.id_artist = app_kpop_group.id
         WHERE FIND_IN_SET(members, ?) AND dead = "n" AND publishedon >= "?-01-01" AND vtype = "main"
         ORDER BY app_kpop.views DESC LIMIT ?;`;
-        db.query(query, [gameSession.getSQLGender(), gameSession.getBeginningCutoffYear(), gameSession.getLimit()])
+        db.query(query, [guildPreference.getSQLGender(), guildPreference.getBeginningCutoffYear(), guildPreference.getLimit()])
         .then((result) => {
             let random = result[Math.floor(Math.random() * result.length)];
             gameSession.startRound(random.name, random.artist, random.youtubeLink);
-            playSong(gameSession, message);
+            playSong(gameSession, guildPreference, db, message);
         })
         .catch((err) => {
             console.log(err);
@@ -85,7 +85,7 @@ module.exports = {
         return message.member.voiceChannel.members.size - 1;
     }
 }
-const playSong = (gameSession, message) => {
+const playSong = (gameSession, guildPreference, db, message) => {
     let voiceChannel = message.member.voiceChannel;
     const streamOptions = { volume: 0.1 };
     voiceChannel.join().then(connection => {
@@ -96,6 +96,6 @@ const playSong = (gameSession, message) => {
         console.error(err);
         // Attempt to restart game with different song
         gameSession.endRound();
-        startGame(message);
+        startGame(gameSession, guildPreference, db, message);
     })
 }
